+++ src/main/java/org/sbml4j/service/NetworkMappingServiceImpl.java
@@ -48,7 +48,7 @@ import org.sbml4j.model.sbml.simple.SBMLSimpleReaction;
 import org.sbml4j.model.sbml.simple.ext.qual.SBMLSimpleTransition;
 import org.sbml4j.model.warehouse.MappingNode;
 import org.sbml4j.model.warehouse.PathwayNode;
-import org.sbml4j.repository.common.BiomodelsQualifierRepository;
+import org.sbml4j.repository.sbml.ext.sbml4j.BiomodelsQualifierRepository;
 import org.sbml4j.service.SimpleSBML.SBMLQualSpeciesService;
 import org.sbml4j.service.SimpleSBML.SBMLSimpleTransitionService;
 import org.sbml4j.service.SimpleSBML.SBMLSpeciesService;
